node-labeller: Decouple host capability inference
This commit introduces a preliminary phase where capability inference
logic is separated from the node-labeller core logic.

The logic is handled by a new component called NodeCapabilitiesManager
that does (for the most part) all of the heavy lifting which was
previously assigned to node-labeller during its construction.

The NodeCapabilitiesManager is instantiated during virt-handler startup
and is passed as a dependency to be used by the node-labeller.

Future commits should strive to decouple the logic further by making
capability management its own separate package within virt-handler.

Signed-off-by: Adi Aloni <adialoni101@gmail.com>