WORKSPACE: Bump grpc-go version add googleapis/rpc
Due to how gazelle works [1] any dependency defined after
gazelle_dependencies() is called that is part of its default dependency
will not override the default.

One of the default gazelle dependencies is grpc-go [2] and therefore
it's required that the dependency definition be stated prior to the
gazelle_dependencies() call.

As a result of bumping grpc-go to v1.65.0 googleapis/rpc is an added
dependency of grpc/status (specified in its BUILD.bazel file).

[1] https://github.com/bazelbuild/bazel-gazelle/issues/1233#issuecomment-1094334092
[2] https://github.com/bazelbuild/bazel-gazelle/blob/v0.33.0/deps.bzl#L258-L264

Signed-off-by: Adi Aloni <adialoni101@gmail.com>