Remove loadHostCapabilities from nodelabeller
The `loadHostCapabilities` function currently unmarshals the entire
`capabilities.xml`, even though the nodelabeller only needs the Host CPU
Counter. THis was necessary because, until now, the virt-handler
depended on the removed `HostCapabilities` function to initialize the
vmcontroller.

This commit changes that by loading the capabilities during the
virt-handler startup and passing the relevant fields to the nodelabeller
and vmcontroller separately. This accomplishes two things:

1. It enhances the testability of the nodelabeller, as it no longer
   needs to simulate file reads to obtain the necessary information
   during tests.

2. It decouples the vmcontroller from the nodelabeller, which was
   previously required to retrieve host capabilities for its
   initialization.

Signed-off-by: Adi Aloni <adialoni101@gmail.com>