Refactor standard vmi-create-admitter tests
This commit is the first in a series of commits to refactor the
vmi-create-admitter tests to not test implementation details and to
uniformly make use of libvmi as the VMI instantiator.

The commit adapts all tests that weren't part of a specific Ginkgo Context
and replaces tests that were testing implementation details in favor of
ones that don't (e.g. "path validation" that tested the validatePath
unexported function instead of calling Admit).

The commit also introduces a new function
"newAdmissionReviewForVMICreation" to uniformly create admission review
objects for VMI creation requests.

Signed-off-by: Adi Aloni <adialoni101@gmail.com>