converter, manager: SetOptimalIOMode dynamic preallocation check
The SetOptimalIOMode used to call the exported function IsPreAllocated
which directly interacts with the filesystem, making it difficult to
unit test.

This commit adds a parameter to SetOptimalIOMode to allow passing a
preallocation checker function and substitute it for unit tests.
Also removed error return value from SetOptimalIOMode as all return
paths were returning nils.

Signed-off-by: Adi Aloni <aaloni@redhat.com>