launcher, converter: Introduce ClockDomainConfigurator
Extract part of the logic responsible for configuring the clock to a
dedicated `ClockDomainConfigurator`.

Remove the existing unit test of `Convert_v1_Clock_To_api_Clock` in
favor of a new test exercising the `ClockDomainConfigurator`.
Add a new test case to test a scenario of a VMI without a clock
specification.
Additional unit tests will be added in a follow-up PR.

Note: The dependencies of `Convert_v1_Clock_To_api_Clock` were
temporarily copied and are planned for refactoring in a follow-up PR.

Signed-off-by: Orel Misan <omisan@redhat.com>