labeller: reduce goroutine count
Currently, the node labeller spawn 10 threads.
All these are watching the same queue.
The queue is populated with the hostname key,
which is unique.
This means that there will always be only one
thread that is doing the job, while the others are
simply doing nothing.

Also, each thread is wrapped by the `wait.Until`.
The wait.Until executes the function f, after it ends,
it waits for the passed interval, and then respawn the
function f.
In this case, since the function f never ends (a forever loop),
the wait.Until is never triggered.
That can simply be replaced by a goroutine.

Signed-off-by: fossedihelm <ffossemo@redhat.com>