config: avoid triggering callbacks on new registering
Currently, every time a ConfigModifiedCallback is registered,
all the callbacks in the slices are triggered.

This is quite inefficient, since the first registered callback
will be executed n-times.
Also, considering that some callbacks performs API Get request, such
as `refreshMediatedDeviceTypes`, this could create api pressure.

Simply execute the callback is going to be registered is quite enough.

Signed-off-by: fossedihelm <ffossemo@redhat.com>