fix(virt-controller): Skip ControllerRevisionRef capture if Matchers nil
As set out in bug #16071 virt-controller would previously panic if a
user removed an instance type or preference from the spec from an
existing VM before restarting it to force a new VM ControllerRevision to
be captured.

This fix avoids the panic for now by skipping the attempted capture of
instance type and preference ControllerRevisionRefs if the associated
spec fields are already nil.

A follow up fix will also handle clearing of these
ControllerRevisionRefs from status when the spec fields are initially
removed from the VM.

Signed-off-by: Lee Yarwood <lyarwood@redhat.com>