e2e,sriov: Reorder operations
We see that some of the tests fail to login to fedora.
The console shows guest is in initial boot sequence (initramfs dracut).

Common reason for those tests is that they don't check agent connected event,
or check it after connecting to console.
This doesn't give enough time to the machine to boot properly.

Reorder to first check agent connected, and then to console, as most of the tests already doing.
This PR address the SR-IOV instance.

https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_kubevirt/15981/pull-kubevirt-e2e-kind-sriov/1992493320420462592

Signed-off-by: Or Shoval <oshoval@redhat.com>