export: introduce export sources
Previously vmexport could only support volume-based export types and the
various existing sources were tightly coupled to the controller logic.

This significantly refactors the controller composition by introducing
the exportSource interface that is fulfilled by all existing export
sources (i.e., pvc, vm and vmsnapshot) to provide the controller
the required context for export creation and handling.

The exportSource interface includes the following methods:
- IsSourceAvailable - Similar to the previously existing sourceVolumes' isSourceAvailable which returns a boolean value of whether the source is available for export.
- HasContent - Returns a boolean value of whether the source has available
content to be exported, for volume-based export types this refers to
whether there are any volumes available for export.
- SourceCondition - Same as sourceVolumes' sourceCondition.
- ReadyCondition - Same as sourceVolumes' ready condition.
- ConfigurePod - Extends the existing createExporterPodManifest controller
method by configuring the pod manifest according to the source.
- ConfigureExportLink - Extends the existing getLinks controller method by
configuring the VirtualMachineExportLink according to the source.
- UpdateStatus - Used by the newly introduced updateStatus controller
  method to update the VirtualMachineExportStatus according to the
  source.

Signed-off-by: Adi Aloni <aaloni@redhat.com>